Requiring login | @login_required

+++ todo/views.py
@@ -5,7 +5,7 @@ from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.shortcuts import get_object_or_404
 from django.contrib import messages
-# Create your views here.
+from django.contrib.auth.decorators import login_required
 
 def get_showing_todos(request, todos):
 
@@ -16,6 +16,8 @@ def get_showing_todos(request, todos):
             return todos.filter(is_completed=False)
     return todos
 
+
+@login_required
 def index(request):
     todos = Todo.objects.filter(owner=request.user)
 
@@ -29,6 +31,8 @@ def index(request):
                 'incomplete_count':incomplete_count}
     return render(request, 'todo/index.html', context)
 
+
+@login_required
 def create_todo(request):
     form = TodoForm()
     context = {'form': form}
@@ -55,26 +59,31 @@ def create_todo(request):
     return render(request, 'todo/create-todo.html', context)
 
 
+@login_required
 def todo_detail(request, id):
     todo = get_object_or_404(Todo, pk=id)
     context = {'todo': todo}
     return render(request, 'todo/todo-detail.html', context)
 
 
+@login_required
 def todo_delete(request, id):
     todo = get_object_or_404(Todo, pk=id)
     context = {'todo': todo}
 
     if request.method == "POST":
-        todo.delete()
+        if todo.owner == request.user:
+            todo.delete()
 
-        messages.add_message(request, messages.SUCCESS, "Todo Deleted Sucessfully")
+            messages.add_message(request, messages.SUCCESS, "Todo Deleted Sucessfully")
 
-        return HttpResponseRedirect(reverse('home'))
+            return HttpResponseRedirect(reverse('home'))
+        return render(request, 'todo/todo-delete.html', context)
 
     return render(request, 'todo/todo-delete.html', context)
 
 
+@login_required
 def todo_edit(request, id):
     todo = get_object_or_404(Todo, pk=id)
     form = TodoForm(instance=todo)
@@ -90,7 +99,8 @@ def todo_edit(request, id):
         todo.description = description
         todo.is_completed = True if is_completed=='on' else False
 
-        todo.save()
+        if todo.owner == request.user:
+            todo.save()
 
         messages.add_message(request, messages.SUCCESS, "Todo Updated Sucessfully")
